Completely re-organize workspace
