+++ .github/workflows/ci.yml
@@ -27,10 +27,10 @@ jobs:
           sudo apt-get install -y --no-install-recommends liblua5.4-dev
 
       - name: Check formatting
-        run: cargo fmt --check
+        run: cargo fmt --check --all
 
       - name: Check Clippy
-        run: cargo clippy --all-targets --features wasm,lua54,bench
+        run: cargo clippy --workspace --all-targets --features wasm,lua54,bench
 
   build:
     runs-on: ${{ matrix.os }}
