+++ benches/serde.rs
@@ -7,13 +7,13 @@ macro_rules! generate_benches {
         $(
             paste! {
                 fn [<corn_ $test_name>](_: ()) -> bool {
-                    let input = include_str!(concat!("../../assets/inputs/", stringify!($test_name), ".corn"));
+                    let input = include_str!(concat!("../assets/inputs/", stringify!($test_name), ".corn"));
                     let output = corn::from_str::<$test_type>(&input);
                     output.is_ok()
                 }
 
                 fn [<json_ $test_name>](_: ()) -> bool {
-                    let input = include_str!(concat!("../../assets/outputs/json/", stringify!($test_name), ".json"));
+                    let input = include_str!(concat!("../assets/outputs/json/", stringify!($test_name), ".json"));
                     let output = serde_json::from_str::<$test_type>(&input);
                     output.is_ok()
                 }
