+++ cli/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "corn-cli"
+version = "0.10.0"
+edition = "2021"
+license = "MIT"
+description = "CLI for Corn. A simple and pain-free configuration language."
+repository = "https://github.com/corn-config/corn"
+categories = ["config", "command-line-utilities"]
+keywords = ["configuration", "language", "pest", "peg", "cli"]
+authors = ["Jake Stanger <mail@jakestanger.com>"]
+readme = "README.md"
+homepage = "https://cornlang.dev/"
+documentation = "https://docs.rs/corn-cli"
+
+[dependencies]
+# CLI dependencies
+clap = { version = "4.5.31", features = ["derive", "cargo"] }
+colored = "3.0.0"
+
+# Deserialization/serialization
+serde = { version = "1.0.218", features = ["derive"] }
+serde_json = "1.0.140"
+serde_norway = "0.9.42"
+libcorn = { version = "0.10.0", path = ".." }
+toml_edit = { version = "0.22.24", features = ["serde"] }
