--- corn-cli/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "corn-cli"
-version = "0.10.0"
-edition = "2021"
-license = "MIT"
-description = "CLI for Corn, A simple and pain-free configuration language."
-repository = "https://github.com/JakeStanger/corn"
-categories = ["config"]
-keywords = ["configuration", "language", "pest", "peg", "cli"]
-
-[dependencies]
-libcorn = { version = "0.10.0", path = "../libcorn" }
-clap = { version = "4.5.31", features = ["derive"] }
-colored = "3.0.0"
-serde = { version = "1.0.218", features = ["derive"] }
-serde_json = "1.0.140"
-serde_norway = "0.9.42"
-toml_edit = { version = "0.22.24", features = ["serde"] }
-pest = "2.7.15"
