--- libcorn/Cargo.toml
@@ -1,61 +1,0 @@
-[package]
-name = "libcorn"
-version = "0.10.0"
-edition = "2021"
-license = "MIT"
-description = "Parsing engine for Corn, a simple and pain-free configuration language."
-repository = "https://github.com/JakeStanger/corn"
-categories = ["config"]
-keywords = ["configuration", "language", "wasm", "pest", "peg"]
-
-[features]
-bench = ["criterion"]
-wasm = [
-    "wasm-bindgen",
-    "serde-wasm-bindgen",
-    "console_error_panic_hook",
-    "wee_alloc",
-]
-lua51 = ["mlua/lua51"]
-lua52 = ["mlua/lua52"]
-lua53 = ["mlua/lua53"]
-lua54 = ["mlua/lua54"]
-luajit = ["mlua/luajit"]
-luajit52 = ["mlua/luajit52"]
-
-[lib]
-name = "corn"
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-pest = "2.7.15"
-pest_derive = "2.7.15"
-cfg-if = "1.0.0"
-serde = { version = "1.0.218", features = ["derive"] }
-indexmap = { version = "2.7.1", features = ["serde"] }
-thiserror = "2.0.12"
-wasm-bindgen = { version = "0.2.100", optional = true }
-serde-wasm-bindgen = { version = "0.6.5", optional = true }
-console_error_panic_hook = { version = "0.1.7", optional = true }
-wee_alloc = { version = "0.4.5", optional = true }
-mlua = { version = "0.10.3", features = [
-    "module",
-    "macros",
-    "serialize",
-], optional = true }
-
-# bench
-criterion = { version = "0.5.1", features = ["html_reports"], optional = true }
-
-[dev-dependencies]
-paste = "1.0.15"
-wasm-bindgen-test = { version = "0.3.50" }
-# required for testing
-serde_json = "1.0.140"
-serde_norway = "0.9.42"
-serde_bytes = "0.11.16"
-toml_edit = { version = "0.22.24", features = ["serde"] }
-
-[[bench]]
-name = "serde"
-harness = false
