--- libcorn/README.md
@@ -1,5 +1,0 @@
-This is the crate for `libcorn`, the parser library for the Corn configuration language.
-
-For the main documentation, check the [repo readme](https://github.com/jakestanger/corn/blob/master/README.md).
-
-For parser API documentation, check [docs.rs](https://docs.rs/libcorn/latest/libcorn/).
